refactor(frontend): split api-keys settings page into components
