+++ frontend/src/components/settings/index.ts
@@ -1,3 +1,9 @@
 export { MigrationBanner } from './MigrationBanner'
 export { EmbeddingModelChangeDialog } from './EmbeddingModelChangeDialog'
 export { ModelTestResultDialog } from './ModelTestResultDialog'
+export { CredentialFormDialog } from './CredentialFormDialog'
+export { DiscoverModelsDialog } from './DiscoverModelsDialog'
+export { DeleteCredentialDialog } from './DeleteCredentialDialog'
+export { CredentialItem } from './CredentialItem'
+export { ProviderSection } from './ProviderSection'
+export { DefaultModelSelectors } from './DefaultModelSelectors'
